site: Refactor, cleanup and add redux
