+++ site/src/app/context.js
@@ -1,0 +1,7 @@
+
+import React from 'react';
+
+const PlaygroundSiteContext = React.createContext();
+
+
+export {PlaygroundSiteContext};
